reuse DirectoryValidator
